backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "MCP Server for Filesystem Operations"
authors = ["Shivam <shivamdave2903@gmail.com>"]
package-mode = false
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = ">=0.115.14,<0.116.0"
uvicorn = ">=0.35.0,<0.36.0"
pydantic = ">=2.0.0,<3.0.0"
pydantic-settings = ">=2.0.0,<3.0.0"
aiofiles = ">=23.0.0,<24.0.0"
python-multipart = ">=0.0.6,<1.0.0"
pathlib2 = ">=2.3.7,<3.0.0"
typing-extensions = ">=4.0.0,<5.0.0"
python-dotenv = ">=1.0.0,<2.0.0"
httpx = ">=0.25.0,<1.0.0"
tenacity = ">=8.0.0,<9.0.0"
together = "^1.5.17"

[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.35.0"}

[tool.poetry.scripts]
start = "uvicorn main:app --reload"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
